refactor: move hedera network utils to correct place
