+++ libs/coin-modules/coin-hedera/src/network/utils.test.ts
@@ -1,9 +1,14 @@
 import { setupMockCryptoAssetsStore } from "@ledgerhq/cryptoassets/cal-client/test-helpers";
+import { InvalidAddress } from "@ledgerhq/errors";
 import { getEnv } from "@ledgerhq/live-env";
 import BigNumber from "bignumber.js";
 import { SUPPORTED_ERC20_TOKENS } from "../constants";
+import { HederaRecipientInvalidChecksum } from "../errors";
 import { getMockedAccount } from "../test/fixtures/account.fixture";
-import { getMockedERC20TokenCurrency } from "../test/fixtures/currency.fixture";
+import {
+  getMockedERC20TokenCurrency,
+  getMockedHTSTokenCurrency,
+} from "../test/fixtures/currency.fixture";
 import {
   getMockedERC20TokenBalance,
   getMockedERC20TokenTransfer,
@@ -13,13 +18,17 @@ import {
   createMirrorTokenTransfer,
   getMockedMirrorTransaction,
   getMockedMirrorContractCallResult,
+  getMockedMirrorAccount,
 } from "../test/fixtures/mirror.fixture";
 import { getMockedConfig } from "../test/fixtures/config.fixture";
 import { getMockedThirdwebTransaction } from "../test/fixtures/thirdweb.fixture";
-import type { HederaMirrorCoinTransfer } from "../types";
+import type { HederaMirrorCoinTransfer, HederaMirrorTransaction } from "../types";
 import { apiClient } from "./api";
 import { hgraphClient } from "./hgraph";
 import {
+  analyzeStakingOperation,
+  calculateUncommittedBalanceChange,
+  checkAccountTokenAssociationStatus,
   createTransactionId,
   enrichERC20Transfers,
   getERC20BalancesForAccount,
@@ -27,6 +36,8 @@ import {
   getERC20Operations,
   parseThirdwebTransactionParams,
   parseTransfers,
+  safeParseAccountId,
+  toEVMAddress,
 } from "./utils";
 
 jest.mock("./api");
@@ -600,6 +611,409 @@ describe("network utils", () => {
       expect(result).toEqual([]);
       expect(apiClient.getContractCallResult).not.toHaveBeenCalled();
       expect(apiClient.findTransactionByContractCallV2).not.toHaveBeenCalled();
+    });
+  });
+
+  describe("checkAccountTokenAssociationStatus", () => {
+    const accountId = "0.0.1234";
+    const htsToken = getMockedHTSTokenCurrency({ contractAddress: "0.0.1234", tokenType: "hts" });
+    const erc20Token = getMockedHTSTokenCurrency({
+      contractAddress: "0.0.4321",
+      tokenType: "erc20",
+    });
+
+    beforeEach(() => {
+      jest.clearAllMocks();
+      // reset LRU cache to make sure all tests receive correct mocks from mockedGetAccount
+      checkAccountTokenAssociationStatus.clear(`${accountId}-${htsToken.contractAddress}`);
+    });
+
+    it("returns true if max_automatic_token_associations === -1", async () => {
+      (apiClient.getAccount as jest.Mock).mockResolvedValueOnce({
+        account: accountId,
+        max_automatic_token_associations: -1,
+        balance: {
+          balance: 0,
+          timestamp: "",
+          tokens: [],
+        },
+      });
+
+      const result = await checkAccountTokenAssociationStatus(accountId, htsToken);
+      expect(result).toBe(true);
+    });
+
+    it("returns true if token is already associated", async () => {
+      (apiClient.getAccount as jest.Mock).mockResolvedValueOnce({
+        account: accountId,
+        max_automatic_token_associations: 0,
+        balance: {
+          balance: 1,
+          timestamp: "",
+          tokens: [{ token_id: htsToken.contractAddress, balance: 1 }],
+        },
+      });
+
+      const result = await checkAccountTokenAssociationStatus(accountId, htsToken);
+      expect(result).toBe(true);
+    });
+
+    it("returns false if token is not associated", async () => {
+      (apiClient.getAccount as jest.Mock).mockResolvedValueOnce({
+        account: accountId,
+        max_automatic_token_associations: 0,
+        balance: {
+          balance: 1,
+          timestamp: "",
+          tokens: [{ token_id: "0.1234", balance: 1 }],
+        },
+      });
+
+      const result = await checkAccountTokenAssociationStatus(accountId, htsToken);
+      expect(result).toBe(false);
+    });
+
+    it("returns true for erc20 tokens", async () => {
+      const result = await checkAccountTokenAssociationStatus(accountId, erc20Token);
+      expect(apiClient.getAccount as jest.Mock).not.toHaveBeenCalled();
+      expect(result).toBe(true);
+    });
+
+    it("supports addresses with checksum", async () => {
+      const addressWithChecksum = "0.0.9124531-xrxlv";
+
+      (apiClient.getAccount as jest.Mock).mockResolvedValueOnce({
+        account: accountId,
+        max_automatic_token_associations: 0,
+        balance: {
+          balance: 1,
+          timestamp: "",
+          tokens: [{ token_id: htsToken.contractAddress, balance: 1 }],
+        },
+      });
+
+      await checkAccountTokenAssociationStatus(addressWithChecksum, htsToken);
+      expect(apiClient.getAccount).toHaveBeenCalledTimes(1);
+      expect(apiClient.getAccount).toHaveBeenCalledWith("0.0.9124531");
+    });
+  });
+
+  describe("safeParseAccountId", () => {
+    it("returns account id and no checksum for valid address without checksum", async () => {
+      const [error, result] = await safeParseAccountId("0.0.9124531");
+
+      expect(error).toBeNull();
+      expect(result?.accountId).toBe("0.0.9124531");
+      expect(result?.checksum).toBeNull();
+    });
+
+    it("returns account id and checksum for valid address with correct checksum", async () => {
+      const [error, result] = await safeParseAccountId("0.0.9124531-xrxlv");
+
+      expect(error).toBeNull();
+      expect(result?.accountId).toBe("0.0.9124531");
+      expect(result?.checksum).toBe("xrxlv");
+    });
+
+    it("returns error for valid address with incorrect checksum", async () => {
+      const [error, accountId] = await safeParseAccountId("0.0.9124531-invld");
+
+      expect(error).toBeInstanceOf(HederaRecipientInvalidChecksum);
+      expect(accountId).toBeNull();
+    });
+
+    it("returns error for invalid address format", async () => {
+      const [error, accountId] = await safeParseAccountId("not-a-valid-address");
+
+      expect(error).toBeInstanceOf(InvalidAddress);
+      expect(accountId).toBeNull();
+    });
+  });
+
+  describe("toEVMAddress", () => {
+    const mockMirrorAccount = {
+      account: "0.0.12345",
+      evm_address: "0x0000000000000000000000000000000000003039",
+    };
+
+    it("returns correct EVM address for valid Hedera account ID", async () => {
+      (apiClient.getAccount as jest.Mock).mockResolvedValueOnce(mockMirrorAccount);
+
+      const evmAddress = await toEVMAddress(mockMirrorAccount.account);
+
+      expect(apiClient.getAccount).toHaveBeenCalledTimes(1);
+      expect(apiClient.getAccount).toHaveBeenCalledWith(mockMirrorAccount.account);
+      expect(evmAddress).toBe(mockMirrorAccount.evm_address);
+    });
+
+    it("returns null when API call fails", async () => {
+      (apiClient.getAccount as jest.Mock).mockRejectedValueOnce(new Error("API error"));
+
+      const evmAddress = await toEVMAddress(mockMirrorAccount.account);
+
+      expect(apiClient.getAccount).toHaveBeenCalledTimes(1);
+      expect(evmAddress).toBeNull();
+    });
+  });
+
+  describe("calculateUncommittedBalanceChange", () => {
+    const mockAddress = "0.0.12345";
+    const mockStartTimestamp = "1762202064.065172388";
+    const mockEndTimestamp = "1762202074.065172388";
+
+    beforeEach(() => {
+      jest.clearAllMocks();
+    });
+
+    it("should return 0 when there are no transactions in the time range", async () => {
+      (apiClient.getTransactionsByTimestampRange as jest.Mock).mockResolvedValueOnce([]);
+
+      const result = await calculateUncommittedBalanceChange({
+        address: mockAddress,
+        startTimestamp: mockStartTimestamp,
+        endTimestamp: mockEndTimestamp,
+      });
+
+      expect(result).toEqual(new BigNumber(0));
+      expect(apiClient.getTransactionsByTimestampRange).toHaveBeenCalledTimes(1);
+      expect(apiClient.getTransactionsByTimestampRange).toHaveBeenCalledWith({
+        address: mockAddress,
+        startTimestamp: `gt:${mockStartTimestamp}`,
+        endTimestamp: `lte:${mockEndTimestamp}`,
+      });
+    });
+
+    it("should calculate balance change with mixed incoming and outgoing transfers", async () => {
+      const mockTransactions = [
+        {
+          consensus_timestamp: "1762202065.000000000",
+          transfers: [
+            { account: mockAddress, amount: 2000 },
+            { account: "0.0.98", amount: -2000 },
+          ],
+        },
+        {
+          consensus_timestamp: "1762202070.000000000",
+          transfers: [
+            { account: mockAddress, amount: -500 },
+            { account: "0.0.99", amount: 500 },
+          ],
+        },
+        {
+          consensus_timestamp: "1762202072.000000000",
+          transfers: [
+            { account: mockAddress, amount: 300 },
+            { account: "0.0.100", amount: -300 },
+          ],
+        },
+      ] as HederaMirrorTransaction[];
+
+      (apiClient.getTransactionsByTimestampRange as jest.Mock).mockResolvedValueOnce(
+        mockTransactions,
+      );
+
+      const result = await calculateUncommittedBalanceChange({
+        address: mockAddress,
+        startTimestamp: mockStartTimestamp,
+        endTimestamp: mockEndTimestamp,
+      });
+
+      expect(result).toEqual(new BigNumber(1800)); // 2000 - 500 + 300
+    });
+
+    it("should ignore transfers for other accounts", async () => {
+      const mockTransactions = [
+        {
+          consensus_timestamp: "1762202065.000000000",
+          transfers: [
+            { account: "0.0.98", amount: 5000 },
+            { account: "0.0.99", amount: -5000 },
+          ],
+        },
+        {
+          consensus_timestamp: "1762202070.000000000",
+          transfers: [
+            { account: mockAddress, amount: 1000 },
+            { account: "0.0.100", amount: -1000 },
+          ],
+        },
+      ] as HederaMirrorTransaction[];
+
+      (apiClient.getTransactionsByTimestampRange as jest.Mock).mockResolvedValueOnce(
+        mockTransactions,
+      );
+
+      const result = await calculateUncommittedBalanceChange({
+        address: mockAddress,
+        startTimestamp: mockStartTimestamp,
+        endTimestamp: mockEndTimestamp,
+      });
+
+      expect(result).toEqual(new BigNumber(1000));
+    });
+
+    it("should return 0 when timestamps are equal or invalid", async () => {
+      (apiClient.getTransactionsByTimestampRange as jest.Mock).mockResolvedValueOnce([]);
+
+      const [resultEqual, resultInvalid] = await Promise.all([
+        calculateUncommittedBalanceChange({
+          address: mockAddress,
+          startTimestamp: mockStartTimestamp,
+          endTimestamp: mockStartTimestamp,
+        }),
+        calculateUncommittedBalanceChange({
+          address: mockAddress,
+          startTimestamp: mockEndTimestamp,
+          endTimestamp: mockStartTimestamp,
+        }),
+      ]);
+
+      expect(resultEqual).toEqual(new BigNumber(0));
+      expect(resultInvalid).toEqual(new BigNumber(0));
+    });
+  });
+
+  describe("analyzeStakingOperation", () => {
+    const mockAddress = "0.0.12345";
+    const mockTimestamp = "1762202064.065172388";
+    const mockTx = {
+      consensus_timestamp: mockTimestamp,
+      name: "CRYPTOUPDATEACCOUNT",
+    } as HederaMirrorTransaction;
+
+    beforeEach(() => {
+      jest.resetAllMocks();
+    });
+
+    it("detects DELEGATE operation when staking starts", async () => {
+      const accountBefore = getMockedMirrorAccount({ staked_node_id: null });
+      const accountAfter = getMockedMirrorAccount({ staked_node_id: 5 });
+
+      (apiClient.getTransactionsByTimestampRange as jest.Mock).mockResolvedValueOnce([]);
+      (apiClient.getAccount as jest.Mock)
+        .mockResolvedValueOnce(accountBefore)
+        .mockResolvedValueOnce(accountAfter);
+
+      const result = await analyzeStakingOperation(mockAddress, mockTx);
+
+      expect(result).toEqual({
+        operationType: "DELEGATE",
+        previousStakingNodeId: null,
+        targetStakingNodeId: 5,
+        stakedAmount: BigInt(1000),
+      });
+      expect(apiClient.getAccount).toHaveBeenCalledTimes(2);
+      expect(apiClient.getAccount).toHaveBeenCalledWith(mockAddress, `lt:${mockTimestamp}`);
+      expect(apiClient.getAccount).toHaveBeenCalledWith(mockAddress, `eq:${mockTimestamp}`);
+    });
+
+    it("detects UNDELEGATE operation when staking stops", async () => {
+      const accountBefore = getMockedMirrorAccount({ staked_node_id: 5 });
+      const accountAfter = getMockedMirrorAccount({ staked_node_id: null });
+
+      (apiClient.getTransactionsByTimestampRange as jest.Mock).mockResolvedValueOnce([]);
+      (apiClient.getAccount as jest.Mock)
+        .mockResolvedValueOnce(accountBefore)
+        .mockResolvedValueOnce(accountAfter);
+
+      const result = await analyzeStakingOperation(mockAddress, mockTx);
+
+      expect(result).toEqual({
+        operationType: "UNDELEGATE",
+        previousStakingNodeId: 5,
+        targetStakingNodeId: null,
+        stakedAmount: BigInt(1000),
+      });
+    });
+
+    it("detects REDELEGATE operation when changing nodes", async () => {
+      const accountBefore = getMockedMirrorAccount({ staked_node_id: 3 });
+      const accountAfter = getMockedMirrorAccount({ staked_node_id: 10 });
+
+      (apiClient.getTransactionsByTimestampRange as jest.Mock).mockResolvedValueOnce([]);
+      (apiClient.getAccount as jest.Mock)
+        .mockResolvedValueOnce(accountBefore)
+        .mockResolvedValueOnce(accountAfter);
+
+      const result = await analyzeStakingOperation(mockAddress, mockTx);
+
+      expect(result).toEqual({
+        operationType: "REDELEGATE",
+        previousStakingNodeId: 3,
+        targetStakingNodeId: 10,
+        stakedAmount: BigInt(1000),
+      });
+    });
+
+    it("calculates correct staked amount with uncommitted transactions", async () => {
+      const mockBalance = { balance: 1000000, timestamp: "1762202060.000000000", tokens: [] };
+      const mockAccountBefore = getMockedMirrorAccount({
+        account: mockAddress,
+        staked_node_id: null,
+        balance: mockBalance,
+      });
+      const mockAccountAfter = getMockedMirrorAccount({
+        account: mockAddress,
+        staked_node_id: 5,
+        balance: mockBalance,
+      });
+      const mockTransactionsMissingInBalance = [
+        {
+          consensus_timestamp: `${Math.floor(Number(mockBalance.timestamp)) + 5}.000000000`,
+          transfers: [
+            { account: mockAddress, amount: -100000 },
+            { account: "0.0.98", amount: 100000 },
+          ],
+        },
+      ] as HederaMirrorTransaction[];
+
+      (apiClient.getTransactionsByTimestampRange as jest.Mock).mockResolvedValueOnce(
+        mockTransactionsMissingInBalance,
+      );
+      (apiClient.getAccount as jest.Mock)
+        .mockResolvedValueOnce(mockAccountBefore)
+        .mockResolvedValueOnce(mockAccountAfter);
+
+      const result = await analyzeStakingOperation(mockAddress, mockTx);
+
+      expect(apiClient.getTransactionsByTimestampRange).toHaveBeenCalledTimes(1);
+      expect(apiClient.getTransactionsByTimestampRange).toHaveBeenCalledWith({
+        address: mockAddress,
+        startTimestamp: `gt:${mockAccountBefore.balance.timestamp}`,
+        endTimestamp: `lte:${mockTimestamp}`,
+      });
+      expect(result).toEqual({
+        operationType: "DELEGATE",
+        previousStakingNodeId: null,
+        targetStakingNodeId: 5,
+        stakedAmount: BigInt(900000),
+      });
+    });
+
+    it("returns null for regular account update (both null)", async () => {
+      const accountBefore = getMockedMirrorAccount({ staked_node_id: null });
+      const accountAfter = getMockedMirrorAccount({ staked_node_id: null });
+
+      (apiClient.getAccount as jest.Mock)
+        .mockResolvedValueOnce(accountBefore)
+        .mockResolvedValueOnce(accountAfter);
+
+      const result = await analyzeStakingOperation(mockAddress, mockTx);
+
+      expect(result).toBeNull();
+    });
+
+    it("returns null when staked node doesn't change", async () => {
+      const accountBefore = getMockedMirrorAccount({ staked_node_id: 5 });
+      const accountAfter = getMockedMirrorAccount({ staked_node_id: 5 });
+
+      (apiClient.getAccount as jest.Mock)
+        .mockResolvedValueOnce(accountBefore)
+        .mockResolvedValueOnce(accountAfter);
+
+      const result = await analyzeStakingOperation(mockAddress, mockTx);
+
+      expect(result).toBeNull();
     });
   });
 });
